Move tcp base entity to separate module
