+++ tests/components/tcp/test_binary_sensor.py
@@ -23,9 +23,9 @@ TEST_ENTITY = "binary_sensor.test_name"
 def mock_socket_fixture():
     """Mock the socket."""
     with (
-        patch("homeassistant.components.tcp.common.socket.socket") as mock_socket,
+        patch("homeassistant.components.tcp.entity.socket.socket") as mock_socket,
         patch(
-            "homeassistant.components.tcp.common.select.select",
+            "homeassistant.components.tcp.entity.select.select",
             return_value=(True, False, False),
         ),
     ):
